# Conflicts:
#	api-version-update/api/models/Suborder.js

+++ api-version-update/api/models/ProductVariant.js
@@ -30,7 +30,7 @@ module.exports = {
     },
     quantity: {
       type: 'number',
-      columnType: 'integer',
+      columnType: 'decimal',
       required: false,
       allowNull: true
     },
